Controllers Auth implemented and working

+++ controllers/profileController.js
@@ -16,9 +16,7 @@ const updateProfile = async (req, res) => {
         const profilePicture = file;
         let bodyObject = {};
 
-        console.log(profilePicture);
-
-        if(JSON.stringify(profilePicture) !== JSON.stringify(oldFileName)) {
+        if(profilePicture && JSON.stringify(profilePicture) !== JSON.stringify(oldFileName)) {
             await unlink(`./images/${oldFileName}`, (err) => {
                 if (err) throw err;
                 console.log('successfully deleted picture');
